\name{capeAddons-package}
\alias{capeAddons-package}
\alias{capeAddons}
\docType{package}
\title{ Support scripts for cape package
}
\description{
Some additional plotting and analysis scripts that work work the cape data object
}
\details{
\tabular{ll}{
Package: \tab capeAddons\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-18\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Maintainer: Anna Tyler <anna.tyler@jax.org>
}
\references{

}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<cape>:<cape>-package]{<cape>}} ~~
}
\examples{
test <- rnorm(100)
}
